requirements.txt
matplotlib==3.9.2
Pillow==9.3.0

